fix: add cron  to send reminders
