+++ .env.example
@@ -6,7 +6,6 @@ HE4RT_TOKEN=sEkM...
 HE4RT_URL=https://...
 
 FIREBASE_DATABASE_URL=https://...
-TENOR_API_KEY=sEkM...
 TENOR_URL=https://...
 APOIASE_URL=https://...
 APOIASE_TOKEN=AiMk...
